+++ src/shops/sell/items.rs
@@ -17,8 +17,8 @@ pub async fn sell(
     #[description = "how many to list"] quantity: i64,
     #[description = "description"] description: Option<String>,
 ) -> Result<(), Error> {
-    let uid = ctx.author().id.get() as i64;
-    let cid = ctx.channel_id().get() as i64;
+    let uid = ctx.author().id.get();
+    let cid = ctx.channel_id().get();
 
     if !db_verify_shop(uid, cid, &ctx.data().database).await? {
         ctx.say("this isn't your shop!").await?;
@@ -55,7 +55,7 @@ pub async fn remove(
     #[description = "item name"] name: String,
     #[description = "quantity"] quantity: i64,
 ) -> Result<(), Error> {
-    let cid = ctx.channel_id().get() as i64;
+    let cid = ctx.channel_id().get();
     db_remove_item(cid, name.as_str(), quantity, &ctx.data().database).await?;
     ctx.say(format!("removed {name} x{quantity}")).await?;
     Ok(())
